+++ src/pages/join/JoinPage.jsx
@@ -1,38 +1,17 @@
 import styles from '@/styles/join/JoinPage.module.css'
 import Button from '@/components/Button'
 import Logo from '@/components/Logo'
-import { useNavigate } from 'react-router-dom'
 import JoinInput from './components/JoinInput'
 import { useState } from 'react'
-import { getJoin } from '@/apis/axios/getJoin'
+import { useHandleJoinClick } from '@/hooks/eventHandler/useHandleJoinClick'
 
 export default function JoinPage() {
-  const [data, setData] = useState(null)
-  const navigate = useNavigate()
   const [groupInput, setGroupInput] = useState('')
   const [error, setError] = useState('')
+  const handleJoinClickHandler = useHandleJoinClick()
 
-  const errorMsg = '잘못된 입력입니다. 다시 시도해주세요.'
-
-  const handleJoinClick = async () => {
-    if (!groupInput) {
-      return
-    }
-
-    const [_groupName, _tag] = groupInput.split('#')
-    if (!_tag || !_groupName) {
-      setError(errorMsg)
-      return
-    }
-
-    try {
-      const res = await getJoin(_groupName, _tag)
-      setData(res)
-      console.log(res)
-      navigate(`/login?group=${_groupName}&tag=${_tag}`)
-    } catch (e) {
-      setError('데이터를 전송하는데 실패하였습니다.')
-    }
+  const handleJoinClick = () => {
+    handleJoinClickHandler(groupInput, setError)
   }
 
   return (
